+++ samples/star/src/commonMain/kotlin/com/slack/circuit/star/home/HomeScreen.kt
@@ -19,11 +19,14 @@ import androidx.compose.runtime.saveable.rememberSaveableStateHolder
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import com.slack.circuit.backstack.rememberSaveableBackStack
 import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.foundation.CircuitContent
+import com.slack.circuit.foundation.LocalCircuit
 import com.slack.circuit.foundation.NavEvent
+import com.slack.circuit.foundation.navEventNavigator
 import com.slack.circuit.foundation.onNavEvent
+import com.slack.circuit.foundation.rememberPresenter
+import com.slack.circuit.foundation.rememberUi
 import com.slack.circuit.retained.rememberRetained
 import com.slack.circuit.runtime.CircuitUiEvent
 import com.slack.circuit.runtime.CircuitUiState
@@ -85,12 +88,28 @@ fun HomeContent(state: HomeScreen.State, modifier: Modifier = Modifier) {
       }
     },
   ) { paddingValues ->
-    val screen = state.navItems[state.selectedIndex].screen
-    CircuitContent(
-      screen,
-      modifier = Modifier.padding(paddingValues),
-      onNavEvent = { event -> state.eventSink(ChildNav(event)) },
-    )
+    val saveableStateHolder = rememberSaveableStateHolder()
+    val currentScreen = state.navItems[state.selectedIndex].screen
+    saveableStateHolder.SaveableStateProvider(currentScreen) {
+      val circuit = requireNotNull(LocalCircuit.current)
+      val ui = rememberUi(currentScreen, factory = circuit::ui)
+      val presenter =
+        rememberPresenter(
+          currentScreen,
+          navigator =
+            Navigator.navEventNavigator(currentScreen) { event ->
+              state.eventSink(ChildNav(event))
+            },
+          factory = circuit::presenter,
+        )
+
+      CircuitContent(
+        screen = currentScreen,
+        modifier = Modifier.padding(paddingValues),
+        presenter = presenter!!,
+        ui = ui!!,
+      )
+    }
     contentComposed = true
   }
   Platform.ReportDrawnWhen { contentComposed }
